# server/src/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Digital-E Shop API",
        "description": "REST API reference for the Digital-E e-commerce backend. Covers catalog browsing, authentication, cart, orders, wishlist, reviews, address book, notifications, inventory, promotions, analytics, and admin operations.",
        "version": "1.0.0",
        "contact": {
            "name": "Digital-E",
            "url": "https://github.com/anomalyco/digital-e-shop"
        },
        "license": {
            "name": "ISC"
        }
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Local development"
        },
        {
            "url": "https://e-commerce-express-server-app.vercel.app",
            "description": "Production"
        }
    ],
...
    "components": {
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "msg": { "type": "string", "description": "Human-readable error message" },
                    "error": { "type": "string", "description": "Machine-readable error code" }
                }
            },
...
            "CartItem": {
                "type": "object",
                "properties": {
                    "cartItemId": { "type": "integer" },
                    "productId": { "type": "integer" },
                    "productName": { "type": "string" },
                    "category": { "type": "string" },
                    "brand": { "type": "string" },
                    "price": { "type": "number" },
                    "sale_price": { "type": "number", "nullable": true },
                    "main_image": { "type": "string" },
                    "quantity": { "type": "integer" },
                    "stock": { "type": "integer" }
                }
            },
            "CartValidationIssue": {
                "type": "object",
                "properties": {
                    "cartItemId": { "type": "integer" },
                    "productId": { "type": "integer" },
                    "reason": { "type": "string", "enum": ["unavailable", "out_of_stock", "insufficient_stock"] },
                    "availableStock": { "type": "integer" }
                }
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "accessToken",
                "description": "JWT access token stored in the accessToken HTTP-only cookie."
            }
        }
    },
    "paths": {
...
        "/api/cart/{uid}": {
            "parameters": [{ "$ref": "#/components/parameters/PathUid" }],
            "get": {
                "tags": ["Cart"],
                "summary": "Get cart for a user (owner or admin)",
                "security": [{ "cookieAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Cart contents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "cartItems": { "type": "array", "items": { "$ref": "#/components/schemas/CartItem" } }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart/{uid}/validation": {
            "parameters": [{ "$ref": "#/components/parameters/PathUid" }],
            "get": {
                "tags": ["Cart"],
                "summary": "Validate cart for checkout (owner or admin)",
                "description": "Returns any products that are out of stock, unavailable, or have insufficient stock.",
                "security": [{ "cookieAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "issues": { "type": "array", "items": { "$ref": "#/components/schemas/CartValidationIssue" } }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cart": {
            "post": {
                "tags": ["Cart"],
                "summary": "Add item to cart (owner or admin)",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["pid", "uid", "quantity"],
                                "properties": {
                                    "pid": { "type": "integer" },
                                    "uid": { "type": "string" },
                                    "quantity": { "type": "integer", "minimum": 1 }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "$ref": "#/components/responses/Message" },
                    "400": { "$ref": "#/components/responses/Error" }
                }
            },
            "put": {
                "tags": ["Cart"],
                "summary": "Update cart item quantity (owner or admin)",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["cartItemId", "quantity"],
                                "properties": {
                                    "cartItemId": { "type": "integer" },
                                    "uid": { "type": "string" },
                                    "quantity": { "type": "integer", "minimum": 1 }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "$ref": "#/components/responses/Message" }
                }
            },
            "delete": {
                "tags": ["Cart"],
                "summary": "Delete cart items (authenticated)",
                "security": [{ "cookieAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "cartItemId": { "oneOf": [{ "type": "integer" }, { "type": "array", "items": { "type": "integer" } }] },
                                    "uid": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "$ref": "#/components/responses/Message" }
                }
            }
        },
...
    },
    "components": {
...
        "parameters": {
...
            "PathUid": {
                "name": "uid",
                "in": "path",
                "required": true,
                "schema": { "type": "string" },
                "description": "Firebase UID"
            }
        },
        "responses": {
            "Message": {
                "description": "Standard message response",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": { "msg": { "type": "string" } }
                        }
                    }
                }
            },
            "Error": {
                "description": "Error response",
                "content": {
                    "application/json": {
                        "schema": { "$ref": "#/components/schemas/Error" }
                    }
                }
            }
        }
    }
}